requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
fake-useragent==1.1.1
idna==3.4
importlib-resources==5.10.2
Pillow==9.4.0
pyTelegramBotAPI==4.4.0
requests==2.28.2
soupsieve==2.3.2.post1
urllib3==1.26.14
vk-api==11.9.5
zipp==3.11.0
